fix!: Removed duplicate graph implementation
